FEATURE: remove environment parsing from launcher.
Rather than parsing the environment yaml config files prior to executing the
container, this moves the responsibility to pups instead which is
more flexible and easier to maintain.

There is an additional benefit that secrets shared via env variables
will not be visible on the container or persisted in any image saved
with `docker commit`.